+++ assets/javascripts/kitten/components/box/link-box/test.js
@@ -1,8 +1,9 @@
 import React from 'react'
 import renderer from 'react-test-renderer'
 import { LinkBox } from './index'
+import { FlashCircleIcon } from '../../icons/flash-circle-icon'
 
-describe('<LinkBox />', () => {
+describe('Deprecated <LinkBox />', () => {
   let component
 
   describe('by default', () => {
@@ -26,6 +27,40 @@ describe('<LinkBox />', () => {
             href="#foobar"
             isExternal
           />,
+        )
+        .toJSON()
+    })
+
+    it('matches with snapshot', () => {
+      expect(component).toMatchSnapshot()
+    })
+  })
+})
+
+describe('<LinkBox />', () => {
+  let component
+
+  describe('by default', () => {
+    beforeEach(() => {
+      component = renderer.create(<LinkBox />).toJSON()
+    })
+
+    it('matches with snapshot', () => {
+      expect(component).toMatchSnapshot()
+    })
+  })
+
+  describe('with some props', () => {
+    beforeEach(() => {
+      component = renderer
+        .create(
+          <LinkBox href="#foobar" isExternal>
+            <LinkBox.Icon>
+              <FlashCircleIcon />
+            </LinkBox.Icon>
+
+            <LinkBox.Text>foobar</LinkBox.Text>
+          </LinkBox>,
         )
         .toJSON()
     })